Divided search_context into each class
